PostItemController, MainMarketView, PostItemService, MarketPlace
Cleaned up the create method in PostItemController, encapsulating some
of the data field assigments that don't get filled by the create view
into a new class, PostItemService.  Changed the initial call to create
in PostItemController so that it no longer takes in a string parameter,
instead dynamically getting the user information.  Changed the call to
this create method in MainMarketView.  Removed unnecessary call to
CNQuery in SetList method in MarketPlace.cs